requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pytest>=7.0
requests>=2.31.0
tenacity>=8.2.0
pydantic[email]>=2.5.0
